Mitigate bad error messages when a template is malformed.
The hand-modifiable templates exposed in the Admin UI expect a fixed number of string fields for substitution.
If the template is modified and a different number of '%s' fields are present, an error is raised.

Before this change, the error message is generic and the stack trace is not very clear.
Furthermore, the error bubbles to the API layer, and is sent to the front-end, where it gets wrongly interpreted as a form field validation error.

This change wraps the error in a more descriptive message, hopefully mitigating this issue.